refactor: move NewBinaryAnalyzer to binary_analyzer.go
NewBinaryAnalyzer is a platform-specific factory for BinaryAnalyzer and
is unrelated to NetworkAnalyzer logic. Moving it to its own file clarifies
the separation between binary format selection and network analysis.
